requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
Flask==3.0.0
flask-cors==4.0.0
idna==3.11
numpy==2.2.6
pandas==2.3.3
pyalex==0.18
python-dotenv==1.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.5.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0

# RAG Chatbot dependencies
langchain-core
langchain
langchain-community
langchain-groq
chromadb
python-dotenv
sentence-transformers
torch

# Flask API dependencies (Milestone 7)
flask
flask-cors